requirements.txt
scikit-learn
Flask
Pandas
Numpy
matplotlib
Seaborn
Pickle
